Replacing short* to float*. We are not running with some STM32 memory constraints or simular theese days.

+++ eurorack/clouds/dsp/pvoc/phase_vocoder.cc
@@ -59,7 +59,7 @@ void PhaseVocoder::Init(
   size_t num_textures = kMaxNumTextures;
   size_t texture_size = (fft_size >> 1) - kHighFrequencyTruncation;
   for (int32_t i = 0; i < num_channels_; ++i) {
-    short* ana_syn_buffer = allocator[i]->Allocate<short>(
+    float* ana_syn_buffer = allocator[i]->Allocate<float>(
         (fft_size + (fft_size >> 1)) * 2);
     
     num_textures = min(
